Added targets and dependencies
Added a Cluster API `Cluster` target to the `AddonConfig` as a means of
getting the resource(s) that the individual `AddonConfig` is associated
with. This `Cluster` serves as one of the first default templating
resources that can be used to help identify `AddonConfigDefinition`
dependencies at runtime.

Added a set of dependencies to the `AddonConfigDefinition` that can be
used to introduce arbitrary resources. These resources are resolved at
runtime by the Addon Config controller using the resolved
`AddonConfig.spec.target` for templating and are then made available for
templating based upon the dependency's given name.

Refactored the controller into phases to help clean up and make clear
the intent of each step.